+++ src/main/java/reactor/core/publisher/FluxSample.java
@@ -53,7 +53,7 @@ final class FluxSample<T, U> extends FluxSource<T, T> {
 	}
 
 	@Override
-	public long getPrefetch() {
+	public int getPrefetch() {
 		return Integer.MAX_VALUE;
 	}
